+++ shared/collision/shape/circle.h
@@ -7,16 +7,16 @@ namespace sunlight::collision
     {
     public:
         Circle() = default;
-        Circle(Eigen::Vector2d center, double radius);
+        Circle(Eigen::Vector2f center, float radius);
 
-        auto GetCenter() const -> Eigen::Vector2d;
-        auto GetRadius() const -> double;
+        auto GetCenter() const -> Eigen::Vector2f;
+        auto GetRadius() const -> float;
 
-        void SetCenter(const Eigen::Vector2d& center);
-        void SetRadius(double radius);
+        void SetCenter(const Eigen::Vector2f& center);
+        void SetRadius(float radius);
 
     private:
-        Eigen::Vector2d _center = {};
-        double _radius = 0.0;
+        Eigen::Vector2f _center = {};
+        float _radius = 0.f;
     };
 }
